Rename logging.py to logger.py to avoid false-positive LGTM alerts
