+++ metaphor/common/extractor.py
@@ -11,7 +11,7 @@ from serde.json import from_json
 from serde.yaml import from_yaml
 from smart_open import open
 
-from metaphor.common.logging import get_logger
+from metaphor.common.logger import get_logger
 
 from .api_sink import ApiSink, ApiSinkConfig
 from .file_sink import FileSink, FileSinkConfig
