+++ metaphor/dbt/extractor.py
@@ -31,7 +31,7 @@ from metaphor.models.metadata_change_event import (
 from metaphor.common.entity_id import EntityId, to_virtual_view_entity_id
 from metaphor.common.event_util import EventUtil
 from metaphor.common.extractor import BaseExtractor
-from metaphor.common.logging import get_logger
+from metaphor.common.logger import get_logger
 from metaphor.dbt.config import DbtRunConfig
 
 from .generated.dbt_catalog_v1 import CatalogTable, DbtCatalog
