+++ metaphor/google_directory/extractor.py
@@ -19,7 +19,7 @@ from smart_open import open
 
 from metaphor.common.event_util import EventUtil
 from metaphor.common.extractor import BaseExtractor, RunConfig
-from metaphor.common.logging import get_logger
+from metaphor.common.logger import get_logger
 
 logger = get_logger(__name__)
