+++ metaphor/looker/extractor.py
@@ -24,7 +24,7 @@ from metaphor.models.metadata_change_event import (
 from metaphor.common.entity_id import to_virtual_view_entity_id
 from metaphor.common.event_util import EventUtil
 from metaphor.common.extractor import BaseExtractor
-from metaphor.common.logging import get_logger
+from metaphor.common.logger import get_logger
 from metaphor.looker.config import LookerConnectionConfig, LookerRunConfig
 from metaphor.looker.lookml_parser import Model, fullname, parse_project
