+++ metaphor/redshift/extractor.py
@@ -3,7 +3,7 @@ from typing import List
 from metaphor.models.metadata_change_event import DataPlatform, MetadataChangeEvent
 
 from metaphor.common.event_util import EventUtil
-from metaphor.common.logging import get_logger
+from metaphor.common.logger import get_logger
 from metaphor.postgresql.extractor import PostgreSQLExtractor
 from metaphor.redshift.config import RedshiftRunConfig
