+++ metaphor/snowflake/extractor.py
@@ -5,7 +5,7 @@ from snowflake.connector import SnowflakeConnection
 from snowflake.connector.cursor import DictCursor, SnowflakeCursor
 
 from metaphor.common.event_util import EventUtil
-from metaphor.common.logging import get_logger
+from metaphor.common.logger import get_logger
 from metaphor.snowflake.auth import connect
 from metaphor.snowflake.config import SnowflakeRunConfig
 from metaphor.snowflake.filter import SnowflakeFilter
