+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "metaphor-connectors"
-version = "0.8.14"
+version = "0.8.15"
 description = ""
 authors = ["Metaphor <dev@metaphor.io>"]
 packages = [
